Moved all *Callback types into /callbacks.rs
