+++ azul/src/widgets/table_view.rs
@@ -2,12 +2,13 @@
 
 use std::collections::BTreeMap;
 use {
-    dom::{Dom, On, NodeData, DomString, NodeType, IFrameCallback, UpdateScreen, DontRedraw},
-    app::AppStateNoData,
     traits::Layout,
+    app::AppStateNoData,
+    callbacks::{IFrameCallback, HidpiAdjustedBounds, UpdateScreen, DontRedraw},
+    dom::{Dom, On, NodeData, DomString, NodeType},
     callbacks::{LayoutInfo, CallbackInfo},
     callbacks::{StackCheckedPointer, DefaultCallback},
-    window::{HidpiAdjustedBounds, FakeWindow},
+    window::FakeWindow,
 };
 
 #[derive(Debug, Default, Copy, Clone)]
